+++ changelog.md
@@ -21,6 +21,7 @@
   - Mapper:
     - Zone/level not updating when active room 
   - Verbatim was failing to append new line characters
+  - Profile manger: Fix display of html characters in tree view
 
 ## 0.4.14 2017-09-04